change points into decimal type

+++ src/public/js/app.js
@@ -76,7 +76,7 @@ const handleFormSubmit = async (event) => {
                 'Content-Type': 'application/json'
             },
             body: JSON.stringify({ 
-                points: parseInt(points),
+                points: parseFloat(points),
                 description
             })
         });
